editoast: move InfraObjects into editoast_schemas
