+++ editoast/src/views/infra/auto_fixes/speed_section.rs
@@ -12,9 +12,9 @@ use crate::generated_data::infra_error::InfraError;
 use crate::generated_data::infra_error::InfraErrorType;
 use crate::infra_cache::operation::CacheOperation;
 use crate::infra_cache::operation::DeleteOperation;
-use crate::infra_cache::operation::InfraObject;
 use crate::infra_cache::operation::Operation;
 use crate::infra_cache::operation::UpdateOperation;
+use editoast_schemas::infra::InfraObject;
 use editoast_schemas::infra::SpeedSection;
 use editoast_schemas::primitives::OSRDIdentified as _;
 use editoast_schemas::primitives::OSRDObject as _;
