+++ editoast/src/views/infra/auto_fixes/track_section.rs
@@ -8,9 +8,9 @@ use super::Fix;
 use crate::generated_data::infra_error::InfraError;
 use crate::generated_data::infra_error::InfraErrorType;
 use crate::infra_cache::object_cache::TrackSectionCache;
-use crate::infra_cache::operation::InfraObject;
 use editoast_schemas::infra::BufferStop;
 use editoast_schemas::infra::Endpoint;
+use editoast_schemas::infra::InfraObject;
 use editoast_schemas::primitives::Identifier;
 use editoast_schemas::primitives::OSRDIdentified as _;
 use editoast_schemas::primitives::OSRDObject as _;
